git-svn-id: http://127.0.0.1/svn/trunk@345 02e793a8-47cb-4219-9cc1-7115cfedab66

+++ src/starcorp/common/entities/ABaseEntity.java
@@ -36,7 +36,7 @@ public abstract class ABaseEntity implements IEntity, Comparable<ABaseEntity> {
 	 * @see starcorp.common.entities.IEntity#readXML(org.dom4j.Element)
 	 */
 	public void readXML(Element e) {
-		this.ID = Integer.parseInt(e.attributeValue("ID","0"));
+		this.ID = Long.parseLong(e.attributeValue("ID","0"));
 	}
 	
 	/* (non-Javadoc)
